# Improve performance of some of the Configurable methods