--- a/include/rocksdb/convenience.h
+++ b/include/rocksdb/convenience.h
@@ -79,6 +79,10 @@ struct ConfigOptions {
   // the value equals the default value.
   bool only_changed_options = false;
 
+  // If true, a best-efforts attempt will be made to restore the original
+  // values of the options when a Configure method fails.
+  bool restore_on_error = true;
+  
   // The separator between options when converting to a string
   std::string delimiter = ";";