--- a/options/options_helper.cc
+++ b/options/options_helper.cc
@@ -729,6 +729,8 @@ Status GetColumnFamilyOptionsFromMap(
     ColumnFamilyOptions* new_options) {
   ConfigOptions copy = config_options;
   copy.only_changed_options = true;
+  copy.restore_on_error = false;
+  
   return GetColumnFamilyOptionsFromMap(copy, ColumnFamilyOptions(), opts_map,
                                        new_options);
 }
@@ -808,6 +810,7 @@ Status GetDBOptionsFromMap(
     DBOptions* new_options) {
   ConfigOptions copy = config_options;
   copy.only_changed_options = true;
+  copy.restore_on_error = false;
   return GetDBOptionsFromMap(copy, DBOptions(), opts_map, new_options);
 }