… table/block_based/block_based_table_factory.cc
… table/plain/plain_table_factory.cc
… tools/dump/rocksdb_dump.cc