This change was motivated to speed up the addition/removal of ColumnFamilies.

The change consists of the following main components:
1.  Some of the internal ConfigurableHelper methods were changed to take/return a vector of string pairs, rather than a map.  Vectors are faster for traversal and removal.
2. The ability to add default values was added to the Configurable and OptionTypeInfo registration.  An option to not print unchanged values was added to the ConfigOptions class.  By not writing the default values, we spent less time in StringToMap and in configuring options that were not changed.
3. New GetXXOptionsFromYY were added that take no default DB/ColumnFamilyOption.  This signature allows the code to know that it does not need to worry about possibly restoring the options to their initial state on failure conditions.  An option was added to ConfigOptions for this instruction as well.
4. If paranoid_checks is off, when an Options file is written, it is not read back in for validation purposes.

With these changes, the time to add/drop 200 ColumnFamilies dropped from 93s to 75s.  10000 calls to the new GetDBOptionsFromMap/GetColumnFamilyOptionsFromMap functions take a 1700 and 2800 ms, respectively vs 2300/3900 ms for the old functions and 3000/6400 in the original code.

Note that this PR does NOT change the options file to only print out the changed options (those that do not match the default).  Doing so would potentially speed up the options file parsing even further but could lead to semantic changes in the future.  A quick test shows the time for Create/DropColumnFamilies dropping to 69s.

There are other performance improvements possible.  For example, not writing an Options file if options were not changed via SetDBOptions could also potentially save some time.  The StringToMap function is also fairly expensive and could potentially be optimized.