@mrambacher 2023-05-16
------------------------------------------------------
./db_bench --benchmarks=readrandom --threads=1 -report_interval_seconds=1 --num=100000000 --value_size=256 --max_background_jobs=12  --compression_type=NONE --duration=300  --db=/tmp/bench --use_existing_db -disable_auto_compactions  --cache_size=$((1024 * 1024 * 1024 * 8))  --disable_auto_compactions --duration=300 --statistics --report_interval_seconds=1 --cache_index_and_filter_blocks=0 -stats_dump_period_sec=60
DB path: [/tmp/bench]
readrandom   :      95.518 micros/op 10469 ops/sec 300.022 seconds 3140999 operations;    1.7 MB/s (1986514 of 3140999 found)

------------------------------------------------------
./db_bench --benchmarks=readrandom --threads=1 -report_interval_seconds=1 --num=100000000 --value_size=256 --max_background_jobs=12  --compression_type=NONE --duration=300  --db=/tmp/bench --use_existing_db -disable_auto_compactions  --cache_size=$((1024 * 1024 * 1024 * 8))  --disable_auto_compactions --duration=300 --statistics --report_interval_seconds=1 --cache_index_and_filter_blocks=1 -stats_dump_period_sec=60
DB path: [/tmp/bench]
readrandom   :      87.854 micros/op 11382 ops/sec 300.023 seconds 3414999 operations;    1.9 MB/s (2160348 of 3414999 found)
------------------------------------------------------
./db_bench --benchmarks=readrandom --threads=1 -report_interval_seconds=1 --num=100000000 --value_size=256 --max_background_jobs=12  --compression_type=NONE --duration=300  --db=/tmp/bench --use_existing_db -disable_auto_compactions  --cache_size=$((1024 * 1024 * 1024 * 8))  --disable_auto_compactions --duration=300 --statistics --report_interval_seconds=1 --cache_index_and_filter_blocks=1 -stats_dump_period_sec=60  --pinning_policy="id=speedb_scoped_pinning_policy;capacity=3221225472"
9.23BG memory usage
DB path: [/tmp/bench]
readrandom   :      87.218 micros/op 11465 ops/sec 300.030 seconds 3439999 operations;    1.9 MB/s (2174623 of 3439999 found)
------------------------------------------------------
./db_bench --benchmarks=readrandom --threads=1 -report_interval_seconds=1 --num=100000000 --value_size=256 --max_background_jobs=12  --compression_type=NONE --duration=300  --db=/tmp/bench --use_existing_db -disable_auto_compactions  --cache_size=$((1024 * 1024 * 1024 * 8))  --disable_auto_compactions --duration=300 --statistics --report_interval_seconds=1 --cache_index_and_filter_blocks=1 -stats_dump_period_sec=60  --pinning_policy="id=speedb_scoped_pinning_policy;capacity=6442450944"
9.23BG memory usage
DB path: [/tmp/bench]
readrandom   :      99.365 micros/op 10063 ops/sec 300.083 seconds 3019999 operations;    1.6 MB/s (1908000 of 3019999 found)